Create hackerRankCodeChallengeDay14.cpp
Objective:-
The absolute difference between two integers, a and b, is written as |a - b|. The maximum absolute difference between two integers in a set of positive integers, elements, is the largest absolute difference between any two integers in _elements.
The Difference class is started for you in the editor. It has a private integer array (elements) for storing  non-negative integers, and a public integer () for storing the maximum absolute difference.

Task:-
Complete the Difference class by writing the following:
1. A class constructor that takes an array of integers as a parameter and saves it to the  instance variable.
2. A computeDifference method that finds the maximum absolute difference between any 2 numbers in _elements and stores it in the maximumDifference instance variable.

Input Format:-
You are not responsible for reading any input from stdin. The locked Solution class in the editor reads in 2 lines of input. The first line contains N, the size of the elements array. The second line has N space-separated integers that describe the _elements array.

Output Format:-
You are not responsible for printing any output; the Solution class will print the value of the maximumDifference instance variable.

Sample Input:-
STDIN   Function
-----   --------
3       __elements[] size N = 3
1 2 5   __elements = [1, 2, 5]
Sample Output:-
4